Starting to organize code
