--- NyachanServer/tag_anime_thread.json
@@ -1,119 +1,0 @@
-{
-    "postsNumber": 8,
-    "imagesNumber": 4,
-    "usersNumber":9,
-    "pageNumber":1,
-    "subject": "Assunto Principal",
-    "isAlive": true,
-    "id": 1113009912,
-    "body": "40666",
-    "date": "2016-01-02 19:31:00",
-    "tile": "Titulo aqui",
-    "userName": "4N0N1M0US3",
-    "file":[
-        {
-            "size": 250,
-            "name": "eu_voto_eh_no_lobao",
-            "extension": "jpg",
-            "source": "http://www.fundosanimais.com/Imagens/imagens-lobos.jpg",
-            "path": ""
-
-        }
-    ],
-    "tags":[
-        "Anime",
-        "Tecnologia",
-        "História"
-    ],
-    "post":[
-        {
-            "id": 111384958,
-            "body": "Nada ver",
-            "date": "2016-01-02 19:34:00",
-            "tile": "Ceguinho",
-            "userName": "4N0Nasdf1M0US3"
-        },
-        {
-            "id": 111384969,
-            "body": "Eitaaa",
-            "date": "2016-01-02 19:35:00",
-            "tile": "Titulo aquasdfsai",
-            "userName": "4N0Nsdf1M0US3"
-        },
-        {
-            "id": 1113851108,
-            "body": "Eu vi tudo",
-            "date": "2016-01-02 19:32:00",
-            "tile": "Quem ta aí?",
-            "userName": "4N0qwN1M1221e0US3",
-            "file":[
-                {
-                    "size": 250,
-                    "name": "fotinea",
-                    "extension": "jpg",
-                    "source": "http://maxscalla.com.br/wp-content/uploads/2013/09/cego.jpg",
-                    "path": ""
-
-                }
-            ]
-        },
-        {
-            "id": 111385158,
-            "body": "Eitaaa",
-            "date": "2016-01-02 19:33:00",
-            "tile": "Titulo qwaasdqui",
-            "userName": "4N0N12w1Masdf0US3",
-            "file":[
-                {
-                    "size": 250,
-                    "name": "fotinea",
-                    "extension": "gif",
-                    "source": "http://337be24b3dc6fa88319a-7601ef3edefa8ce5b9804dfa0e8536c1.r65.cf2.rackcdn.com/da30ca911754e6a36bd5f0df19a5549d.gif",
-                    "path": ""
-
-                }
-            ]
-        },
-        {
-            "id": 111385168,
-            "body": "Ta tudo errado",
-            "date": "2016-01-02 19:32:00",
-            "tile": "Amadake",
-            "userName": "4N0qwN1M0US3",
-            "file":[
-                {
-                    "size": 250,
-                    "name": "fotinea",
-                    "extension": "jpg",
-                    "source": "https://www.cavzodiaco.com.br/hadessantuario/foto4.jpg",
-                    "path": ""
-
-                }
-            ]
-        },
-        {
-            "id": 1111411,
-            "body": "Eitaaa",
-            "date": "2016-01-02 19:33:00",
-            "tile": "O mesmo de sempre",
-            "userName": "4N0N1Masdf0US3"
-        },
-        {
-            "id": 1111311,
-            "body": "Segundo o/",
-            "date": "2016-01-02 19:32:00",
-            "tile": "Eu",
-            "userName": "4N0qwN1M0US3",
-            "file":[
-                {
-                    "size": 250,
-                    "name": "Mais 1",
-                    "extension": "jpg",
-                    "source": "http://picview.info/download/20160312/saitama-one-punch-man-anime-3840x2160.jpg",
-                    "path": ""
-
-                }
-            ]
-        }
-    ]
-}
